requirements.txt
PyStemmer>=1.1.0
lxml>=2.2.4
feedparser>=4.1
mongoengine>=0.3
